Harden repository security and reliability
