+++ .dockerignore
@@ -1,0 +1,18 @@
+.git
+.github
+.idea
+.venv
+**/__pycache__
+**/.pytest_cache
+**/*.pyc
+.env
+config
+library
+backend/library
+frontend/node_modules
+frontend/dist
+frontend/playwright-report
+frontend/test-results
+test-results
+*.log
+.DS_Store
