+++ .github/workflows/ci.yml
@@ -12,16 +12,15 @@ jobs:
     permissions:
       contents: read
     steps:
-    - uses: actions/checkout@v3
+    - uses: actions/checkout@v4
     - name: Set up Python
-      uses: actions/setup-python@v4
+      uses: actions/setup-python@v5
       with:
         python-version: '3.13.5'
     - name: Install uv
-      run: curl -LsSf https://astral.sh/uv/install.sh | sh
+      uses: astral-sh/setup-uv@v6
     - name: Install dependencies
       run: |
-        source ~/.cargo/env
         uv venv
         source .venv/bin/activate
         uv pip install -e ".[dev]"
@@ -41,16 +40,15 @@ jobs:
   test:
     runs-on: ubuntu-latest
     steps:
-    - uses: actions/checkout@v3
+    - uses: actions/checkout@v4
     - name: Set up Python
-      uses: actions/setup-python@v4
+      uses: actions/setup-python@v5
       with:
         python-version: '3.13.5'
     - name: Install uv
-      run: curl -LsSf https://astral.sh/uv/install.sh | sh
+      uses: astral-sh/setup-uv@v6
     - name: Install dependencies
       run: |
-        source ~/.cargo/env
         uv venv
         source .venv/bin/activate
         uv pip install -e ".[dev]"
@@ -60,6 +58,43 @@ jobs:
         source .venv/bin/activate
         pytest -m "not integration" backend/tests/
 
+  migrations:
+    runs-on: ubuntu-latest
+    permissions:
+      contents: read
+    steps:
+    - uses: actions/checkout@v4
+    - name: Set up Python
+      uses: actions/setup-python@v5
+      with:
+        python-version: '3.13.5'
+    - name: Install uv
+      uses: astral-sh/setup-uv@v6
+    - name: Install dependencies
+      run: |
+        uv venv
+        uv pip install -e ".[dev]"
+    - name: Test PostgreSQL migrations
+      run: make test-migrations
+
+  dependency-audit:
+    runs-on: ubuntu-latest
+    permissions:
+      contents: read
+    steps:
+    - uses: actions/checkout@v4
+    - name: Set up Python
+      uses: actions/setup-python@v5
+      with:
+        python-version: '3.13.5'
+    - name: Install uv
+      uses: astral-sh/setup-uv@v6
+    - name: Audit Python dependencies
+      run: |
+        uv venv
+        uv pip install -e ".[dev]"
+        uvx pip-audit --path .venv/lib/python3.13/site-packages
+
   lint-ui:
     runs-on: ubuntu-latest
     permissions:
@@ -68,7 +103,7 @@ jobs:
       run:
         working-directory: frontend
     steps:
-    - uses: actions/checkout@v3
+    - uses: actions/checkout@v4
     - name: Set up Node.js
       uses: actions/setup-node@v4
       with:
@@ -79,6 +114,10 @@ jobs:
       run: npm ci
     - name: Lint with eslint
       run: npm run lint -- --max-warnings=0
+    - name: Audit production dependencies
+      run: |
+        npm audit --omit=dev --audit-level=moderate
+        npm audit --audit-level=high
 
   ui-tests:
     runs-on: ubuntu-latest
@@ -86,7 +125,7 @@ jobs:
       run:
         working-directory: frontend
     steps:
-    - uses: actions/checkout@v3
+    - uses: actions/checkout@v4
     - name: Set up Node.js
       uses: actions/setup-node@v4
       with:
