+++ backend/app/upload_validation.py
@@ -13,16 +13,23 @@ _ZIP_MAGIC_EMPTY = b"PK\x05\x06"  # empty archive
 MAX_UPLOAD_BYTES = 500 * 1024 * 1024  # 500 MB
 MAX_ZIP_ENTRIES = 5000  # max files inside a ZIP/EPUB
 MAX_ZIP_UNCOMPRESSED_RATIO = 100  # compressed-to-uncompressed ratio (zip bomb detection)
+MAX_ZIP_UNCOMPRESSED_BYTES = 2 * MAX_UPLOAD_BYTES  # cap expanded batch archives at 1 GB
 MAX_IMAGE_BYTES = 10 * 1024 * 1024  # 10 MB for cover images
+UPLOAD_READ_CHUNK_BYTES = 1024 * 1024
+
 
 # Image magic bytes
-_IMAGE_SIGNATURES = {
-    b"\xff\xd8\xff": "JPEG",
-    b"\x89PNG\r\n\x1a\n": "PNG",
-    b"RIFF": "WEBP",  # WEBP starts with RIFF....WEBP
-    b"GIF87a": "GIF",
-    b"GIF89a": "GIF",
-}
+def detect_image_extension(payload: bytes) -> str | None:
+    """Return a safe raster extension based on file contents, never the filename."""
+    if payload.startswith(b"\xff\xd8\xff"):
+        return ".jpg"
+    if payload.startswith(b"\x89PNG\r\n\x1a\n"):
+        return ".png"
+    if payload.startswith((b"GIF87a", b"GIF89a")):
+        return ".gif"
+    if payload.startswith(b"RIFF") and len(payload) >= 12 and payload[8:12] == b"WEBP":
+        return ".webp"
+    return None
 
 
 def validate_magic_bytes(payload: bytes, filename: str) -> None:
@@ -45,7 +52,7 @@ def validate_file_size(payload: bytes, filename: str) -> None:
         size_mb = len(payload) / (1024 * 1024)
         limit_mb = MAX_UPLOAD_BYTES // (1024 * 1024)
         raise HTTPException(
-            status_code=status.HTTP_413_REQUEST_ENTITY_TOO_LARGE,
+            status_code=status.HTTP_413_CONTENT_TOO_LARGE,
             detail=f"Uploaded file '{filename}' is {size_mb:.1f} MB, exceeding the {limit_mb} MB limit.",
         )
 
@@ -63,6 +70,11 @@ def validate_zip_safety(payload: bytes, filename: str) -> None:
 
             compressed_size = sum(e.compress_size for e in entries)
             uncompressed_size = sum(e.file_size for e in entries)
+            if uncompressed_size > MAX_ZIP_UNCOMPRESSED_BYTES:
+                raise HTTPException(
+                    status_code=status.HTTP_400_BAD_REQUEST,
+                    detail=(f"'{filename}' expands to more than " f"{MAX_ZIP_UNCOMPRESSED_BYTES // (1024 * 1024)} MB."),
+                )
             if compressed_size > 0 and uncompressed_size / compressed_size > MAX_ZIP_UNCOMPRESSED_RATIO:
                 raise HTTPException(
                     status_code=status.HTTP_400_BAD_REQUEST,
@@ -91,14 +103,29 @@ def validate_upload(payload: bytes, filename: str) -> None:
     validate_zip_safety(payload, filename)
 
 
+async def read_upload_limited(file: UploadFile, max_bytes: int, filename: str) -> bytes:
+    """Read at most max_bytes from an upload, rejecting before unbounded allocation."""
+    payload = bytearray()
+    while True:
+        chunk = await file.read(min(UPLOAD_READ_CHUNK_BYTES, max_bytes + 1 - len(payload)))
+        if not chunk:
+            return bytes(payload)
+        payload.extend(chunk)
+        if len(payload) > max_bytes:
+            raise HTTPException(
+                status_code=status.HTTP_413_CONTENT_TOO_LARGE,
+                detail=f"Uploaded file '{filename}' exceeds the {max_bytes // (1024 * 1024)} MB limit.",
+            )
+
+
 async def read_and_validate_upload(file: UploadFile) -> bytes:
-    """Read an UploadFile and run all validations. Returns the raw bytes."""
+    """Read an UploadFile within the EPUB limit and run all validations."""
     if not file.filename:
         raise HTTPException(
             status_code=status.HTTP_400_BAD_REQUEST,
             detail="Uploaded file is missing a filename.",
         )
-    payload = await file.read()
+    payload = await read_upload_limited(file, MAX_UPLOAD_BYTES, file.filename)
     validate_upload(payload, file.filename)
     return payload
 
@@ -113,10 +140,10 @@ def validate_image_upload(payload: bytes, filename: str) -> None:
     if len(payload) > MAX_IMAGE_BYTES:
         size_mb = len(payload) / (1024 * 1024)
         raise HTTPException(
-            status_code=status.HTTP_413_REQUEST_ENTITY_TOO_LARGE,
+            status_code=status.HTTP_413_CONTENT_TOO_LARGE,
             detail=f"Image '{filename}' is {size_mb:.1f} MB, exceeding the {MAX_IMAGE_BYTES // (1024 * 1024)} MB limit.",
         )
-    if not any(payload.startswith(sig) for sig in _IMAGE_SIGNATURES):
+    if detect_image_extension(payload) is None:
         raise HTTPException(
             status_code=status.HTTP_400_BAD_REQUEST,
             detail=f"Uploaded file '{filename}' is not a recognized image format (JPEG, PNG, WEBP, GIF).",
