+++ backend/tests/test_upload_validation.py
@@ -8,8 +8,9 @@ from fastapi import HTTPException
 
 from backend.app.upload_validation import (
     MAX_IMAGE_BYTES,
-    MAX_UPLOAD_BYTES,
     MAX_ZIP_ENTRIES,
+    detect_image_extension,
+    read_upload_limited,
     validate_image_upload,
     validate_magic_bytes,
     validate_file_size,
@@ -54,20 +55,36 @@ class TestValidateFileSize:
     def test_within_limit(self):
         validate_file_size(b"x" * 100, "small.epub")  # should not raise
 
-    def test_exceeds_limit(self):
-        # We can't allocate 500MB in tests, so we monkey-patch conceptually.
-        # Instead, test with a payload larger than MAX_UPLOAD_BYTES.
-        # Use a small fake to test the logic by checking the threshold.
-        # Just verify the boundary condition with a smaller payload.
-        payload = b"x" * (MAX_UPLOAD_BYTES + 1)
+    def test_exceeds_limit(self, monkeypatch):
+        monkeypatch.setattr("backend.app.upload_validation.MAX_UPLOAD_BYTES", 16)
+        payload = b"x" * 17
         with pytest.raises(HTTPException) as exc_info:
             validate_file_size(payload, "huge.epub")
         assert exc_info.value.status_code == 413
         assert "exceeding" in exc_info.value.detail
 
-    def test_exactly_at_limit(self):
-        payload = b"x" * MAX_UPLOAD_BYTES
+    def test_exactly_at_limit(self, monkeypatch):
+        monkeypatch.setattr("backend.app.upload_validation.MAX_UPLOAD_BYTES", 16)
+        payload = b"x" * 16
         validate_file_size(payload, "exact.epub")  # should not raise
+
+
+@pytest.mark.asyncio
+async def test_read_upload_limited_stops_after_limit():
+    class ChunkedUpload:
+        def __init__(self):
+            self.read_sizes = []
+
+        async def read(self, size):
+            self.read_sizes.append(size)
+            return b"x" * size
+
+    upload = ChunkedUpload()
+    with pytest.raises(HTTPException) as exc_info:
+        await read_upload_limited(upload, 8, "too-large.epub")
+
+    assert exc_info.value.status_code == 413
+    assert sum(upload.read_sizes) == 9
 
 
 class TestValidateZipSafety:
@@ -109,6 +126,14 @@ class TestValidateZipSafety:
         assert exc_info.value.status_code == 400
         assert "entry limit" in exc_info.value.detail
 
+    def test_rejects_excessive_total_uncompressed_size(self, monkeypatch):
+        monkeypatch.setattr("backend.app.upload_validation.MAX_ZIP_UNCOMPRESSED_BYTES", 4)
+        payload = _make_zip({"large.txt": b"12345"})
+        with pytest.raises(HTTPException) as exc_info:
+            validate_zip_safety(payload, "expanded.zip")
+        assert exc_info.value.status_code == 400
+        assert "expands" in exc_info.value.detail
+
 
 class TestValidateUpload:
     def test_valid_epub_like_zip(self):
@@ -131,6 +156,13 @@ class TestValidateImageUpload:
         payload = b"\x89PNG\r\n\x1a\n" + b"\x00" * 100
         validate_image_upload(payload, "cover.png")  # should not raise
 
+    def test_rejects_svg(self):
+        with pytest.raises(HTTPException):
+            validate_image_upload(b"<svg><script>alert(1)</script></svg>", "cover.svg")
+
+    def test_rejects_non_webp_riff(self):
+        assert detect_image_extension(b"RIFF\x10\x00\x00\x00WAVEfake") is None
+
     def test_empty_image(self):
         with pytest.raises(HTTPException) as exc_info:
             validate_image_upload(b"", "empty.jpg")
@@ -147,3 +179,6 @@ class TestValidateImageUpload:
         with pytest.raises(HTTPException) as exc_info:
             validate_image_upload(payload, "huge.jpg")
         assert exc_info.value.status_code == 413
+
+    detect_image_extension,
+    read_upload_limited,
