+++ docker-compose.e2e.yml
@@ -1,4 +1,3 @@
-version: '3.8'
 services:
   story-manager-e2e:
     image: story-manager-e2e:latest
