+++ docker-compose.yml
@@ -1,4 +1,3 @@
-version: '3.8'
 services:
   story-manager:
     build: .
