+++ pyproject.toml
@@ -18,31 +18,31 @@ dependencies = [
     "cloudscraper==1.2.71",
     "ebooklib==0.20",
     "fanficfare==4.56.0",
-    "fastapi==0.136.0",
+    "fastapi==0.139.0",
     "greenlet==3.4.0",
     "h11==0.16.0",
     "html2text==2025.4.15",
     "html5lib==1.1",
     "httpcore==1.0.9",
     "httptools==0.7.1",
-    "idna==3.11",
-    "lxml==6.0.4",
+    "idna==3.18",
+    "lxml==6.1.1",
     "pydantic==2.13.2",
     "pydantic-core==2.46.2",
     "python-dotenv==1.2.2",
-    "python-multipart==0.0.26",
+    "python-multipart==0.0.32",
     "pyyaml==6.0.3",
     "requests==2.33.1",
     "requests-file==3.0.1",
     "requests-toolbelt==1.0.0",
     "six==1.17.0",
     "sniffio==1.3.1",
-    "soupsieve==2.8.3",
+    "soupsieve==2.8.4",
     "sqlalchemy==2.0.49",
-    "starlette==1.0.0",
+    "starlette==1.3.1",
     "typing-inspection==0.4.2",
     "typing-extensions==4.15.0",
-    "urllib3==2.6.3",
+    "urllib3==2.7.0",
     "uvicorn==0.44.0",
     "uvloop==0.22.1",
     "watchfiles==1.1.1",
@@ -53,7 +53,7 @@ dependencies = [
 [project.optional-dependencies]
 dev = [
     "aiosqlite==0.22.1",
-    "httpx==0.28.1",
+    "httpx2==2.5.0",
     "iniconfig==2.3.0",
     "packaging==26.1",
     "pluggy==1.6.0",
@@ -77,6 +77,7 @@ packages = ["backend"]
 
 [tool.black]
 line-length = 127
+target-version = ["py313"]
 
 [tool.flake8]
 max-line-length = 127
